+++ test/unit/handle-known-defects.test.js
@@ -1,4 +1,4 @@
-const handleKnownDefects = require('../../app/processing/parsing/handle-known-defects')
+const handleKnownDefects = require('../../app/processing/siti-agri/handle-known-defects')
 const { sfiPilot } = require('../../app/schemes')
 const GROSS_LINE_DESCRIPTION = 'G00 - Gross value of claim'
 const PARTICIPATION_PAYMENT_SCHEME_CODE = '80009'
