+++ test/unit/processing/siti-agri/get-value-in-pence.test.js
@@ -1,4 +1,4 @@
-const getValueInPence = require('../../../../app/processing/parsing/get-value-in-pence')
+const getValueInPence = require('../../../../app/processing/siti-agri/get-value-in-pence')
 
 describe('get invoice total', () => {
   test('calculates invoice total if integer single line', () => {
